Separate generators.
